feat: add Compose search UI, multiselect toggle, and Android Compose docs mirror
- PagerScreen: OutlinedTextField search + LaunchedEffect to ViewModel query
- HomeAppItem / AppIcon: wire onToggleSelection callback for multiselect
- PagerViewModel: setQuery + toggleSelection helpers
- PagerFragment + fragment_pager.xml: thread search and selection state
- docs/superpowers/plans/2026-09-02-compose-migration-plan.md: update plan with Phase 1 findings
- .kilo/skills/jetpack-compose/reference: local mirror of Android Compose developer docs
- .kilo/reference/developer_android_compose_llms.txt: bundled LLM-friendly docs dump

+++ app/src/main/kotlin/com/aistra/hail/ui/screens/home/PagerViewModel.kt
@@ -31,7 +31,7 @@ class PagerViewModel(application: Application) : AndroidViewModel(application) {
     private val _selectedList = MutableStateFlow<List<AppInfo>>(emptyList())
     val selectedList: StateFlow<List<AppInfo>> = _selectedList.asStateFlow()
 
-    var selectedTagId: Long = 0
+    var selectedTagId: Int = 0
         private set
 
     init {
@@ -40,11 +40,11 @@ class PagerViewModel(application: Application) : AndroidViewModel(application) {
         }
     }
 
-    fun loadApps(tagId: Long = selectedTagId) {
+    fun loadApps(tagId: Int = selectedTagId) {
         val queryText = _query.value
         val filtered = HailData.checkedList.filter { it.isInstalled }
         val result = if (queryText.isEmpty()) {
-            filtered.filter { it.tagIdList.contains(tagId.toInt()) }
+            filtered.filter { it.tagIdList.contains(tagId) }
         } else {
             filtered.filter { app ->
                 ((HailData.nineKeySearch && NineKeySearch.search(
@@ -83,7 +83,7 @@ class PagerViewModel(application: Application) : AndroidViewModel(application) {
         _multiselect.value = false
     }
 
-    fun setTagId(tagId: Long) {
+    fun setTagId(tagId: Int) {
         selectedTagId = tagId
         loadApps(tagId)
     }
